# BE/swagger-main.yaml
openapi: 3.0.3
info:
  title: Discord Clone API
  description: |
    A REST API for managing channels, direct messages, servers, and users in a Discord clone application.
  version: 1.0.0
servers:
  - url: http://localhost:3000/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Server:
      type: object
      properties:
        id:
          type: string
          example: '617a518f130b3fdd2a5c89a9'
        serverName:
          type: string
          example: 'My Server'
        owner:
          type: string
          example: '614b518f130b3fdd2a5c89c9'
        members:
          type: array
          items:
            type: string
          example: ['614b518f130b3fdd2a5c89c9', '615b518f130b3fdd2a5c89e0']
        imageUrl:
          type: string
          example: 'uploads/serverImage.png'
    ServerCreate:
      type: object
      properties:
        serverName:
          type: string
          example: 'My New Server'
        ownerId:
          type: string
          example: '66f9015f6adef60dff23d15e'
        image:
          type: string
          format: binary
    ServerUpdate:
      type: object
      properties:
        serverName:
          type: string
          example: 'My Updated Server'
        image:
          type: string
          format: binary
...
    Message:
      type: object
      properties:
        _id:
          type: string
          example: '614b518f130b3fdd2a5c89e1'
        senderId:
          type: object
          properties:
            _id:
              type: string
              example: '614b518f130b3fdd2a5c89e0'
            name:
              type: string
              example: 'John Doe'
            imageUrl:
              type: string
              example: 'uploads/userImage.png'
        content:
          type: string
          example: 'Hello, everyone!'
        channelId:
          type: string
          example: '614b518f130b3fdd2a5c89d9'
    MessageCreate:
      type: object
      properties:
        content:
          type: string
          example: 'This is a new message'
    MessageDelete:
      type: object
      properties:
        messageId:
          type: string
          example: '614b518f130b3fdd2a5c89e1'
    DirectMessage:
      type: object
      properties:
        _id:
          type: string
          example: '617a518f130b3fdd2a5c89a9'
        senderId:
          type: object
          properties:
            senderId:
              type: string
              example: '614b518f130b3fdd2a5c89c9'
            name:
              type: string
              example: 'John Doe'
            imageUrl:
              type: string
              example: 'uploads/userImage.png'
        recipientId:
          type: string
          example: '614b518f130b3fdd2a5c89e1'
        message:
          type: string
          example: 'Hello, how are you?'
        timestamp:
          type: string
          format: date-time
          example: '2024-08-23T12:00:00Z'
...
paths:
...
  /server:
    get:
      tags:
        - Server
      summary: Get all servers
      description: Retrieve a list of all servers.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
        '401':
          description: Unauthorized access

    post:
      tags:
        - Server
      summary: Create a new server
      description: Create a new server with an optional image. The **ownerId** is extracted from the JWT token and not provided in the request body.
      security:
        - bearerAuth: []
      requestBody:
        description: Server creation data (the ownerId is extracted from the JWT token)
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '201':
          description: Server created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized access
  /server/{id}:
    get:
      tags:
        - Server
      summary: Get server by ID
      description: Retrieve a specific server's details using its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the server to retrieve
      responses:
        '200':
          description: Server details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '401':
          description: Unauthorized access
        '404':
          description: Server not found

    put:
      tags:
        - Server
      summary: Update server details
      description: Update a server's details such as its name and image.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the server to update
      requestBody:
        description: Server update data
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: Server updated successfully
        '400':
          description: Invalid input
        '401':
          description: Unauthorized access
        '404':
          description: Server not found

    delete:
      tags:
        - Server
      summary: Delete server by ID
      description: Delete a server by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the server to delete
      responses:
        '204':
          description: Server deleted successfully
        '401':
          description: Unauthorized access
        '404':
          description: Server not found
  /server/user/servers:
    get:
      tags:
        - Server
      summary: Get all servers by user
      description: Retrieve a list of all servers the user belongs to.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
        '401':
          description: Unauthorized access
...
  /message/{channelId}:
    get:
      tags:
        - Message
      summary: Get all messages in a channel
      description: Retrieve all messages for a specific channel.
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: ID of the channel to retrieve messages for
      responses:
        '200':
          description: A list of messages for the channel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized access

    post:
      tags:
        - Message
      summary: Create a new message in a channel
      description: Create a new message in a specific channel. The senderId is extracted from the JWT token.
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: ID of the channel to send the message to
      requestBody:
        description: Message data (content only, senderId is extracted from the JWT token)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '201':
          description: Message created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized access

    delete:
      tags:
        - Message
      summary: Delete a message in a channel
      description: Delete a specific message from a channel. The message can only be deleted by the user who sent it.
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: ID of the channel where the message was sent
      requestBody:
        description: Message to be deleted (messageId required)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageDelete'
      responses:
        '200':
          description: Message deleted successfully
        '400':
          description: Invalid input
        '401':
          description: Unauthorized access
...
  /directMessage/{recipientId}:
    get:
      tags:
        - Direct Message
      summary: Get all messages in a conversation with a specific user
      description: Retrieve all direct messages exchanged between the current user and a specific recipient.
      security:
        - bearerAuth: []
      parameters:
        - name: recipientId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the recipient
      responses:
        '200':
          description: A list of messages in the conversation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DirectMessage'
        '401':
          description: Unauthorized access